Update README.md
# This is my frist work!
Basically, this is for you to quickly test whether every data at the left of an excel is bigger than the right. The thing you need to do is introduce the data of excel into Matlab with matrix data and put it in the first code line: a = ' '. For example, if your excel is named "data", what you need to do is introduce it and replace the 'file name' with data, like a= data. And then if every data on the left of an excel is bigger than the right it shows true, otherwise, it will show false. A very simple program so you can modify it into a much more powerful tool.